Todo list application
This is a fully working version of the program using MVC with Maven, Servlets and JSPs. The information is persisted via JDBC code.

The application can gather info from a form, redirects to appropriate JSP page if the information is valid. There is a simple validation.

This program can do all the CRUD operations with Database using MVC Model 2 Architecture.

There are separatate Servlets for each business scenarios.